# Sped up simulator using wyrand PRNG.